機能改善

+++ classes/AppManager.py
@@ -54,6 +54,9 @@ class AppManager(object):
         return True
 
     async def agree(self, message: discord.Message):
+        if message.author.id in self.agree_users:
+            await message.author.send("すでに同意しています。")
+            return
         if message.guild:
             await message.delete()
         await message.author.send("利用規約承認を確認しました。\nSigma を使用することができるようになりました。\n`sigma`でご利用ください。")
